probe_info_service: implement a connector for metadata storage
This CL introduces a connector module that performs IO of
the probe metadata.  The selected data storage is
`google.cloud.datastore`.  In order to perform unittests locally,
the deploy script is also modified so that a local emulator
will be run before the tests.

BUG=b:154137694, b:146768256
TEST=./deploy/probe_info_service.sh unittest

Change-Id: Iaad1cc88e96ff45c981a5cce1630e9d6e74f72a8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/2151367
Commit-Queue: Yong Hong <yhong@chromium.org>
Tested-by: Yong Hong <yhong@chromium.org>
Reviewed-by: Liang-Chieh Chen <jamesqaq@chromium.org>