test_lists: Clear TPM ownership in GRT
Currently most test lists clear TPM ownership in FFT, but this is
error-prone when we try to override the test list (e.g. forget to add
this item). Not clearing TPM ownership leads to finalization failure.

Since we always need to run this for finalization to pass, it makes more
sense to move it to GRTItems.

BUG=None
TEST=None

Change-Id: Ib57f2e051c46a220d624b52af2cfe7738b66b6ef
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/2152259
Reviewed-by: Stimim Chen <stimim@chromium.org>
Reviewed-by: Yilin Yang (kerker) <kerker@chromium.org>
Commit-Queue: Cheng-Han Yang <chenghan@chromium.org>
Tested-by: Cheng-Han Yang <chenghan@chromium.org>