tpm_verify_ek: wait and check TPM is owned before verifying key
When we tries to acquire TPM ownership via `tpm-manager initialize`, the
return code implies if the TakeOwnership command is sent successfully
but not if the initialization is done. To ensure TPM is properly owned
and password is generated, we need to rely on the response of
`cryptohome --action=tpm-status`.
Adding a poll on TPM status after initialization command is sent.

BUG=b:153767839
TEST=manual on Krane

Change-Id: Id5dd2ae66c84489e81ddc766e75fa009ec0d8b80
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/2182091
Reviewed-by: Stimim Chen <stimim@chromium.org>
Commit-Queue: Fei Shao <fshao@chromium.org>
Tested-by: Fei Shao <fshao@chromium.org>