update_cbi: Update project_config from Dome
This CL adds an argument "enable_factory_server" in update_cbi. If it is
set then we download the project_config files from factory server.

This CL also change how the version of project_config is calculated
since gz format includes timestamps of files and the hash value may be
changed even if the content is the same.

BUG=b:152291015
TEST=(1) ./setup/cros_docker.sh build && ./setup/cros_docker.sh run
     (2) Restart the umpire server in DOME UI
     (3) Upload project_config component to DOME
     (4) 'factory run Provision.UpdateSKUID' on DUT
     (5) See the project_config is installed at
     /usr/local/factory/project_config
     (6) 'factory run Provision.UpdateSKUID' on DUT again
     (7) The project_config is not installed twice

Change-Id: I041a068d4192a89e0a39fe659ce9718d01effb58
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/2178805
Reviewed-by: Stimim Chen <stimim@chromium.org>
Commit-Queue: Cheng Yueh <cyueh@chromium.org>
Tested-by: Cheng Yueh <cyueh@chromium.org>