factory: Set encoding for `subprocess.Popen`
`subprocess.Popen` uses bytes by default, we will get errors if
we treat it as str type when processing. In python3, we should use str
type as much as possible, so we need to set the encoding or use
`process_utils.Spawn` to get str type.

BUG=chromium:999876
TEST=make test
TEST=manually run code segment in dut python interpreter

Change-Id: Icf617d1d9f2473f6074623ac230c36a702e51d04
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/2182095
Commit-Queue: Yilin Yang (kerker) <kerker@chromium.org>
Tested-by: Yilin Yang (kerker) <kerker@chromium.org>
Reviewed-by: Pin-yen Lin <treapking@chromium.org>