test_list_checker: Fix `--board` option
When `--board` argument is given, the test_list_checker builds an
overlay directory and run the checks.  Because the code generates
test list directory relative to the factory directory, we need to
use the factory_env in overlay directory for correct factory
directory path.

BUG=None
TEST=test_list_checker --board puff main_puff

Change-Id: I0cb9a64e841679b02c727f3fb6bf25c4c750ec77
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/2206555
Reviewed-by: Stimim Chen <stimim@chromium.org>
Tested-by: Pin-yen Lin <treapking@chromium.org>
Commit-Queue: Pin-yen Lin <treapking@chromium.org>