evdev_utils: Print how devices are filtered in error
Previously if there are none or multiple devices found in
evdev_utils.FindDevice, we did not tell the user which devices are
filtered by which rules and what are the multiple candidates left.

This CL prints those information so we can set the arguments more
easily.

BUG=b:157402556
TEST=factory run FFT.Keyboard
TEST=factory run FFT.LidSwitch

Change-Id: Ia7069c65f23441019c69025ac5c56c2d01044d4a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/2215925
Reviewed-by: Stimim Chen <stimim@chromium.org>
Tested-by: Cheng Yueh <cyueh@chromium.org>
Commit-Queue: Cheng Yueh <cyueh@chromium.org>