multicast: Fix zombie process issue
In the original implementation, if the multicast service is
deactivated, no one will handle the return status of uftp processes,
which leads to many zombie processes in the container.  This CL
handles the SIGTERM signal and manually kill and wait child uftp
processes to prevent zombie processes.

BUG=b:155147650
TEST=Toggle multicast service and check `ps aux | grep uftp`

Change-Id: I0f5acfba0e09e4f144bf2693aa52a0b477cb8641
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/2224461
Reviewed-by: Stimim Chen <stimim@chromium.org>
Commit-Queue: Pin-yen Lin <treapking@chromium.org>
Tested-by: Pin-yen Lin <treapking@chromium.org>