instalog: Fix TypeError in input_http_testlog plugin
When the log level is not provided in Testlog JSON, the default value
was `logging.getLevelName(logging.CRITICAL)`, which is 'CRITICAL' and
cannot be compared with a int object (self.threshold).  Thus a
TypeError exception would be raised.  Also, avoid using
`logging.getLevelName` on string variables.

BUG=None
TEST=Send `station.message` request without specifying log level

Change-Id: I1719f89e1d8801c8128e8573cd4782fed06c85a6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/2223840
Reviewed-by: Chun-Tsen Kuo <chuntsen@chromium.org>
Commit-Queue: Pin-yen Lin <treapking@chromium.org>
Tested-by: Pin-yen Lin <treapking@chromium.org>