wireless_antenna: Fix wifi_chip_type detection
Previously we assume that WiFi chip either support switching (tx, rx) to
(1, 1), (2, 2), and (3, 3) or none of them. Sometimes a WiFi chip may
only support a partial of them. This kind of WiFi chips will be tested
in radiotap mode. We need to switch it back to use all antennas.
Otherwise, the chip will be tested in radiotap mode with one antenna
switched off and fail.

BUG=b:158042867
TEST=factory run FFT.Wireless
TEST=run wireless_antenna on DUT with such type of chip

Change-Id: If280b0f70d5f300ee28265ab529d35c70c70b5b6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/2230680
Reviewed-by: Yong Hong <yhong@chromium.org>
Tested-by: Cheng Yueh <cyueh@chromium.org>
Commit-Queue: Cheng Yueh <cyueh@chromium.org>