probe_info_service: use different configs based on the environment
A configuration file in the private repo now provides configuration
sets for different Google Cloud Projects.  Now, during initialization,
the service loads the configuration file and picks up the right
config set based on the name of the GCP.

In addition, this CL pipes python's root logger to google cloud logging
API if the runtime environment is not local.  The major benefit is to
have a more structural logging message on the log viewer.

BUG=b:146768256
TEST=manual test

Cq-Depend: chrome-internal:3104778
Change-Id: Ieeb57e19ab9191b48808c854fa665bae34f164c8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/2244671
Tested-by: Yong Hong <yhong@chromium.org>
Commit-Queue: Yong Hong <yhong@chromium.org>
Reviewed-by: Clark Chung <ckclark@chromium.org>