pylint: Fix comparison-with-callable lint error
comparison-with-callable is a new checker to prevent error-prone code
like:
  return func == 123
which should be:
  return func() == 123

But we do want to compare if the method itself is the default one.

Using `is` for comparison still works because the default sleep
function should be always referring to the same object.

BUG=b:158811392
TEST=make lint && make test
TEST=manual tests on the condition:
     `_GetPollingSleepFunction() is _DEFAULT_POLLING_SLEEP_FUNCTION`
     returns true iff sleep function is the default one

Change-Id: I5f0325e35591a3a3b46a92cb05a8cbfe4ad2955e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/2247563
Reviewed-by: Yilin Yang (kerker) <kerker@chromium.org>
Tested-by: Fei Shao <fshao@chromium.org>
Commit-Queue: Fei Shao <fshao@chromium.org>