net_utils: unify return value of GetEthernetIp
The return value of net_utils.GetEthernetIp(...) could be a single
string (IP address) or a tuple (IP address, prefix number).
Which will be returned depends on the second argument "netmask" is true
or not.
This function prototype is error-prone, for example, when converting the
return value to boolean:

  bool(None) == False
  bool((None, None)) == True

We might hit this "feature" when we change netmask from false to true.
Or even worse, when we change the default value of netmask.

This removed the netmask argument, and make the function always returns
a tuple.  We also changed all callers so the additional return value is
handled properly.

BUG=None
TEST=make test

Change-Id: I141f66f550d495bd1c92f5bfaf0cc7fd8e225ebb
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/2276071
Reviewed-by: Yilin Yang (kerker) <kerker@chromium.org>
Tested-by: Stimim Chen <stimim@chromium.org>
Commit-Queue: Stimim Chen <stimim@chromium.org>