audio_loop: Allow longer duration in sinewav test
Previously, we didn't check if the duration is longer than the wav file.
This CL adjusts the length of the wav file to be always longer than the
duration and fix BaseAudioControl._GetPIDByName if no process presents.

Also shows a more clear log when aplay ends before we call
_StopPlaybackWavFile.

BUG=b:150912730
TEST=(1) `factory run SMT.AudioJack` with duration=20 without this CL
         => test fails
     (2) `factory run SMT.AudioJack` with duration=20 with this CL
         => test passes

Change-Id: I8e968dbb9a24651ca27c4a42549830b1316cba3e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/2879119
Tested-by: Cheng Yueh <cyueh@chromium.org>
Reviewed-by: Chung-Sheng Wu <chungsheng@chromium.org>
Commit-Queue: Cheng Yueh <cyueh@chromium.org>