test: Remove WriteProtectSwitch in generic test lists
The WriteProtectSwitch pytest is designed for testing the write
protect screw in the previous devices.  Now the write protection is
controlled by H1 chip, and this pytest will always fail because the
H1 chip is in factory mode before finalization.  Therefore this CL
removes the test item in generic test lists.  This CL keeps the
definition in generic_common so it don't break other test lists, and
a warning message is added in the WriteProtectSwitch pytest.

BUG=None
TEST=None

Change-Id: I5359964b2a3d36898844b1d3a6c7c03ad249ffa0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/2282624
Reviewed-by: Stimim Chen <stimim@chromium.org>
Tested-by: Pin-yen Lin <treapking@chromium.org>
Commit-Queue: Pin-yen Lin <treapking@chromium.org>