sh: Add private_overlay_presubmit_check tool
This provides the presubmit check for private overlay to check the
health of test lists.

"test_list_checker" needs to be run in chroot when checking the private
overlay. This CL refers to go/croscli/2991495. It tries to run
`cros_sdk` if it's not in chroot.

BUG=b:161196535
TEST=test "repo upload" inside/outside chroot on jacuzzi board

Change-Id: I85c0670b8512dcc4aabb8934ea4e31eb23761d29
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/2297162
Commit-Queue: Yilin Yang (kerker) <kerker@chromium.org>
Tested-by: Yilin Yang (kerker) <kerker@chromium.org>
Reviewed-by: Pin-yen Lin <treapking@chromium.org>