probe_info_service: fix MethodNotFound error on staging service
The entry module mistakenly called the test-purpose methods in the
helper instance.  The behavior turns out being only valid in a local
environment.  This CL fixes the issue by correcting the called method.

In addition, this CL corrects the testdata which is found to be a
legacy snapshot.

BUG=None
TEST=manual test; ./deploy/probe_info_service.sh test

Change-Id: I5cff7e7ca0ddf3242efde09fa4a5467d6caaa4e0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/2309873
Tested-by: Yong Hong <yhong@chromium.org>
Commit-Queue: Yong Hong <yhong@chromium.org>
Reviewed-by: Liang-Chieh Chen <jamesqaq@chromium.org>