test_list_checker: Validate the value of run_if
In the past, there was no unique name to define if a component is
existed or not. For example, zork/Vilboz uses has_keyboard_backlight
and zork/Ezkinil uses has_kbbacklight.

Nowadays, boxster creates the model_sku.json for some projects, and the
names are unified. We need a check to raise the warning when we use the
wrong naming.

Note: We only check keys in this form "component.has_*".

BUG=chromium:782179
TEST=manually run test_list_checker on juniper test list

Change-Id: I1deb53bb167e8c9c3d05a076b31820399245745c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/2324269
Reviewed-by: Cheng Yueh <cyueh@chromium.org>
Commit-Queue: Yilin Yang (kerker) <kerker@chromium.org>
Tested-by: Yilin Yang (kerker) <kerker@chromium.org>