led: Tell operator to change pd status in the UI
On Volteer, there are two leds but the 'ectool led' only let us control
the color but not which led is lit. Which led is lit is controlled by
the usbpdpower status.

This CL add the logics to check the usbpdpower status so we can test
different leds.

BUG=b:162397149
TEST=factory run SMT.LED

Change-Id: Ica3a84bb7df4928c89036c276602ddeb63bbbf23
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/2331787
Reviewed-by: Jian-Jia Su <jjsu@chromium.org>
Tested-by: Cheng Yueh <cyueh@chromium.org>
Commit-Queue: Cheng Yueh <cyueh@chromium.org>