hwid/service: sync HWID DB from git repo
The current mechanism which syncs HWID DB to cloud storage daily will
cause a timing issue if we also cache the latest commit hash to stop the
payload generation.  We cache the hash of commit which will regenerate
the payload but the HWID DB in cloud storage is out-of-date.  HWID
Service will then stop to generate the payload until the HWID DB in
cloud storage and the commit hash are both updated again.

This CL resolves this issue by fetching the up-to-date HWID DB from git
repo instead of the daily sync'ed staging folder of cloud storage.

BUG=b:171032627
TEST=${FACTORY_REPO}/deploy/cros_hwid_service.sh test

Cq-Depend: chrome-internal:3347971
Change-Id: Ia8c4e1601eb2bffe7554f8722622cceae0056651
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/2484067
Commit-Queue: Clark Chung <ckclark@chromium.org>
Tested-by: Clark Chung <ckclark@chromium.org>
Reviewed-by: Yong Hong <yhong@chromium.org>