mrc_cache: Provide mrc_cache tool and refine factory test
In factory, engineers need to retrain memory manually sometimes. But we
don't have a tool to provide such function.

Besides, the retrain process in the original pytests is not accurate:
1. It doesn't retrain RW_MRC_CACHE for x86.
2. It doesn't retrain RO_DDR_TRAINING for ARM (Qualcomm).
3. Actually, we don't need a recovery boot for arm series.

Hence, this CL sorts out the retrain process and provide an independent
tool for direct use.

BUG=b:154782084
TEST=Run MRCCache pytest on soraka: x86
TEST=Run MRCCache pytest on juniper: ARM (Mediatek)

Change-Id: Ifa0706c8a76aad1deb9297673bb34c9a90b064f7
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/2383404
Reviewed-by: Yu-Ping Wu <yupingso@chromium.org>
Reviewed-by: Stimim Chen <stimim@chromium.org>
Reviewed-by: Hung-Te Lin <hungte@chromium.org>
Commit-Queue: Yilin Yang (kerker) <kerker@chromium.org>
Tested-by: Yilin Yang (kerker) <kerker@chromium.org>