keyboard: Fix propertyNames in jsonschema
The attribute 'propertyNames' should be an object or boolean. The
propertyNames doesn't work for jsonschema 2.6.0 and doesn't raise error
if propertyNames is in the schema so the keyboard test can pass with
jsonschema 2.6.0.

This CL fixes propertyNames so the keyboard test can pass with
jsonschema 2.6.0 or jsonschema 3.2.0.

This CL also adds the unittest for this case. Update chroot or
'python3 -m pip install jsonschema==3.2.0' if you fail the unittest.

BUG=None
TEST='factory run SMT.Keyboard' on 13433.0.0 test image
TEST=(1) python3 -m pip install jsonschema==2.6.0
     (2) './py/test/pytests/keyboard_unittest.py' and fail
     (3) python3 -m pip install jsonschema==3.2.0
     (4) './py/test/pytests/keyboard_unittest.py' and pass

Change-Id: I29d9dd0c99b6d45807b78cb9b52e8da824c3248c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/2389285
Reviewed-by: Yilin Yang (kerker) <kerker@chromium.org>
Commit-Queue: Cheng Yueh <cyueh@chromium.org>
Tested-by: Cheng Yueh <cyueh@chromium.org>