time_sanitizer: Change base time if ahead of server time
The system time of the factory servers might not be correct, so
sometimes the base time (usually be the time of OS installation)
might be ahead of the system time on factory server.  Modify the
base time to prevent this from failing reboot tests.

BUG=b:169542240
TEST=manually test on DUT

Change-Id: Ic54cb612231f11bbb06ce54be616c5a8439ebc6c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/2440135
Reviewed-by: Cheng Yueh <cyueh@chromium.org>
Commit-Queue: Pin-yen Lin <treapking@chromium.org>
Tested-by: Pin-yen Lin <treapking@chromium.org>