argparse_utils: Make subparser required in py3
Similar to CL:2157341, subparser is not required by default in py3, and
we will get a confusing error message when the command is not complete.
With this CL, when the command is not complete, it will print out the
help message and a missing argument error message.

BUG=None
TEST=make test; manual run commands such as `gooftool/commands.py`.

Change-Id: I92fe484e5e2a5e0a4cb19c7dc56e2cd21e6e27c5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/2465356
Tested-by: Cheng-Han Yang <chenghan@chromium.org>
Reviewed-by: Yilin Yang (kerker) <kerker@chromium.org>
Commit-Queue: Cheng-Han Yang <chenghan@chromium.org>