ovl: Support specifying certificate path
In the past, overlord used so-called self-signed certificate. When ovl
connect to it, it can download the certificate and use it to verify.
Because it's signed by itself.

However, with https://crrev.com/c/2467736, we create our own root CA and
use it to sign the certificate for overlord. Then it's impossible to
pass the verification if we can't specify the correct certificate.
Hence, `ovl connect` should be able to specify the path to the correct
root CA certificate to verify the connection.

BUG=b:170172074
TEST=Specify correct root CA, and `ovl connect` succeed
TEST=Specify wrong certificate, and `ovl connect` fail

Change-Id: I271354141514473b05e55e7ec19ff2453ef78cf0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/2497638
Reviewed-by: Cheng Yueh <cyueh@chromium.org>
Commit-Queue: Yilin Yang (kerker) <kerker@chromium.org>
Tested-by: Yilin Yang (kerker) <kerker@chromium.org>