overlord: Add TrackConnection flag
We want to track connection even it's disconnected when we want to do
so. For the sake of this, we need a flag stored in overlord server, and
it can be controlled from ghost client.

We'll use this flag in the following CLs.

BUG=b:172784970
TEST=`ghost --track-connection y`
     1. (in dut) print correct log
     2. Use `docker logs {overlord container name} -f` shows correct log

Change-Id: Ic1f5e774ca79d0cb5f888c4d73d2dcf56dcfc985
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/2530866
Reviewed-by: Cheng Yueh <cyueh@chromium.org>
Tested-by: Yilin Yang (kerker) <kerker@chromium.org>
Commit-Queue: Yilin Yang (kerker) <kerker@chromium.org>