overlord: Support Disconnected status
After we set `TrackConnection` and `TimeoutSecs` for a specific Mid
client, the overlord server will keep the connection info. When the
connection of that client is disconnected, the overlord server will wait
for a while until timeout, and then show "Disconnected" status in UI.

Besides, we also need to support when a disconnected client connects
back to overlord server. When a client connects back, we check if the
old connection is disconncted. If yes, we inherit some settings from the
old connection, and don't emit the disconnected event.

BUG=b:172784970
TEST=1. (in dut) `ghost --track y --time 60` start to track the connection
     2. Remove network cable to force disconnected
     3. (in overlord server) `ovl ls` shows disconnected after about 60
     seconds
TEST=1. (in dut) `ghost --track y -- time 60` start to track the connection
     2. Remove network cable to force disconnected
     3. Plug in network cable and set ip
     4. "ghost" automatically connect to overlord server
     5. UI doesn't show disconnected later

Change-Id: I34eb54a921aa58a208bb09d5824546713cb0a36e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/2546250
Tested-by: Yilin Yang (kerker) <kerker@chromium.org>
Commit-Queue: Yilin Yang (kerker) <kerker@chromium.org>
Reviewed-by: Cheng Yueh <cyueh@chromium.org>