goofy_ghost: Run ghost in "goofy_control start"
The current boot flow (See more details in init/README.md):
1. (upstart) factory-init
2. (upstart) boot_services
3. (upstart) factory
4. goofy_control start

We run "ghost" in "factory-init" stage now. In this case, we always get
the wrong test list which is "generic_main". The reason is that we can't
get the correct model name at this time. Because we mount
"/run/chromeos-config/v1" when "boot_services" is going to start. Check
"cros_configfs.conf" for more details.

To solve this problem, we can launch "ghost" in "goofy_control start"
step.

BUG=b:172784970
TEST=1. reboot dut
     2. check "/var/run/factory/goofy_ghost.json", the test list is
     correct

Change-Id: I5390af020418f444f22203cae769c5398d842b59
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/2546249
Reviewed-by: Cheng Yueh <cyueh@chromium.org>
Reviewed-by: Pi-Hsun Shih <pihsun@chromium.org>
Tested-by: Yilin Yang (kerker) <kerker@chromium.org>
Commit-Queue: Yilin Yang (kerker) <kerker@chromium.org>