test_list_checker: Add waived options
We may want to waive some of the warnings in the CQ. For example,
unified device.component.has_* name is a nice to have convention but
voilating it does not break the test lists.

This CL marks checks that break test lists as FATAL or ERROR and adds
waived options to let us only fail the CQ if the change breaks the test
lists.

Message types modified from pylint convention:
  * (N) none, which fails if any check fails
  * (C) convention, for programming standard violation
  * (W) warning, for test list specific problems
  * (E) error, for much probably bugs in the test list
  * (F) fatal, if an error occurred which prevented test_list_checker
        from doing

BUG=b:172526627
TEST=PROGRAM=volteer && \
     LIST_DIR=overlay-${PROGRAM}/py/test/test_lists && \
     make "overlay-${PROGRAM}" && \
     LISTS=$(for x in ${LIST_DIR}/*.test_list.json; do \
       echo $(basename ${x} .test_list.json); \
       done;) && \
     bin/test_list_checker --board ${PROGRAM} ${LISTS} --waived W;\
     echo $?

Change-Id: If00f738f5d680777db0b91a12866c68b6c201fe2
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/2567863
Reviewed-by: Yilin Yang (kerker) <kerker@chromium.org>
Tested-by: Cheng Yueh <cyueh@chromium.org>
Commit-Queue: Nicolas Boichat <drinkcat@chromium.org>