multicast: Run UFTP servers on published host ports
The new multicast container shares the same network with the host
machine, so we cannot set the multicast server port to the default port
inside the container (8080) and we don't need (and cannot do) port
forwarding.
This CL set the multicast default port to external umpire port so that
there shouldn't be collision with other projects. UFTP server uses UDP
so this won't affect the umpire's TCP port.

BUG=b:155147650, b:176517254
TEST=None

Change-Id: I5bd804b20ba129f92c307dbd0b7f84aef84cb186
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/2589177
Tested-by: Pin-yen Lin <treapking@chromium.org>
Commit-Queue: Pin-yen Lin <treapking@chromium.org>
Reviewed-by: Cheng Yueh <cyueh@chromium.org>