factory_bug: Use relative path for zip archive
Currently, the files are collected from the absolute path. If the files
be in the mounted devices, the mount point will be included in the path
of the files in the zip archive.
To make the paths of files always the same, symlinks are created before
collecting logs. All the files now use relative path related to the
temporary directory.

BUG=b:147462001
TEST=manually test on device

Change-Id: I6bae516f455a7b86b774a2619940be3fbc46d319
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/2529792
Tested-by: Chung-Sheng Wu <chungsheng@chromium.org>
Reviewed-by: Fei Shao <fshao@chromium.org>
Commit-Queue: Chung-Sheng Wu <chungsheng@chromium.org>