hwid: rename "video" component to "camera"
Now, the probe statement of "video" should only return camera components
(USB webcam or MIPI camera).  Rename the component name to "camera" so
it is easier to understand.  And "camera" also matches the component
type on DLM.

Note that, we DO NOT expect this change to be backported to factory
branches.

BUG=b:176400434
TEST=make test
TEST=hwid build-database ==> the component name is "camera"

Change-Id: Ia6a33946db23280aa9d35937f78eec23c7c45d01
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/2607916
Tested-by: Stimim Chen <stimim@chromium.org>
Reviewed-by: Yong Hong <yhong@chromium.org>
Commit-Queue: Stimim Chen <stimim@chromium.org>