ghost: Store track_connection info in memory
We want to start to track the connection at certain point in the test
list. But there may be no network at that moment. So we need to keep the
track_connection info in memory, and send it to overlord server when the
connection is constructed.

Since we store the info in memory, so restarting ghost clears the info,
and it won't send any track_connection info to overlord server to
prevent wrong setting.

BUG=b:172784970
TEST=(Assume eth0 is the interface for DUT to connect to overlord
server)
     1. (DUT) `ifconfig eth0 down`
     2. (Overlord server) `ovl ls` doesn't show the DUT
     3. (DUT) `ghost --track y`
     4. (DUT) `ifconfig eth0 up`
     5. (Overlord server) `ovl ls` shows the DUT, and track_connection
     shows true.

Change-Id: I4bb2866291d04e4323d5859875e93f069ccfbe3a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/2607918
Commit-Queue: Yilin Yang (kerker) <kerker@chromium.org>
Tested-by: Yilin Yang (kerker) <kerker@chromium.org>
Reviewed-by: Cheng Yueh <cyueh@chromium.org>