external_display: Support HW buttons when testing
When testing external display on tablet, it's really inconvenient for
ODM to attach a keyboard and then press the digit number. Factory eng
wrote a special version of pytest for this request (e.g. in kukui).

This CL tries to integrate the idea from kukui branch into tot and makes
some appropriate modification.

After this CL, as long as we fill up the hw_buttons and device_filter
(if needed) params, the external_display test will switch to use the HW
buttons to validate.

BUG=b:136784757
TEST=manual run external display on DUT

Change-Id: I2141a671b6e5f428fa58151e0e81613ecad99394
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/2622936
Reviewed-by: Fei Shao <fshao@chromium.org>
Tested-by: Yilin Yang (kerker) <kerker@chromium.org>
Commit-Queue: Yilin Yang (kerker) <kerker@chromium.org>