dome: Rename dome_mroute to dome_mcast and mount volumes
In the new multicast design, we don't need a multicast routing daemon,
but we want to run the multicast server on a host-network container
rather than run in Umpire (which is in bridged network).
This change contains following changes to run the multicast server in a
new container:
(1) Rename `dome_mroute` to `dome_mcast` for multicast server.
(2) Mount Umpire directory to run multicast with Umpire resources.
(3) Mount `MCAST_SHARED_DIR` for multicast logs.
(4) Mount /var/run/docker.sock to communicate with the docker daemon.

BUG=b:155147650, b:176517254
TEST=Build and run Dome

Change-Id: Ic42b82b28533ecdd5c0db7c991b32b93330e32e8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/2635501
Tested-by: Pin-yen Lin <treapking@chromium.org>
Commit-Queue: Pin-yen Lin <treapking@chromium.org>
Reviewed-by: Yilin Yang (kerker) <kerker@chromium.org>
Reviewed-by: Cheng Yueh <cyueh@chromium.org>