overlord: Refine light feature
Light feature: It's a feature that used in FixtureWidget to indicate
something is on or off.

Defect in original design:
1. on/off texts are the same.
2. on/off colors are bright green and bright red.

Combined with the above two facts, it's a little hard for human to aware
the light change.

Here comes to the new design:
1. Change the on/off color to bright green and gray. It's a strong
contrast. And gray is intuitive for human to indicate 'off'.
2. The shown text can changed between on/off color.

You can see picture from http://b/171753480#comment12.

BUG=b:172784970, b:171753480
TEST=manual test on overlord UI

Change-Id: I6c4fcf3335abb59b0b2bebb8ee437455e8ab4e14
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/2631824
Commit-Queue: Yilin Yang (kerker) <kerker@chromium.org>
Tested-by: Yilin Yang (kerker) <kerker@chromium.org>
Reviewed-by: Jian-Jia Su <jjsu@chromium.org>