hwid_extractor: Update config through WebUI
Add interface to update the config. The config is saved on the server
side. Each time the page is reload the config.json is fetched from
server. If the users change the config the change is saved to server
automatically.
Add interface to update RLZ mapping data.

BUG=b:174717120
TEST=make check_js

Change-Id: I8009cf811927888b670d792cffff006f3e690497
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/2652114
Tested-by: Chung-Sheng Wu <chungsheng@chromium.org>
Commit-Queue: Chung-Sheng Wu <chungsheng@chromium.org>
Reviewed-by: Clark Chung <ckclark@chromium.org>
Reviewed-by: Jinsong Mu <jinsong@google.com>