overlord: Support setting port forward in UI
We try to make use of "ghost" and "ovl" to achieve this feature.

First, we run "ghost" in overlord server host machine, then we can
control this host machine through "ghost".

Later, when users click the manage button and open the management UI, we
send a command to host machine to do the following two things:
1. Run "ovl connect" to connect to overlord server
2. Run "ovl forward" to establish the port forward mapping

Then users can click the "Goofy" link from UI to open client's goofy UI.

Finally, when users close the management UI, we remove the port forward
setting.

One special point is that when "ovl" connects to overlord server, we
need to specify the account and password. To achieve this, we create a
special "ovl" account when doing "cros_docker.sh overlord setup". And
use that account / automatic-gen password for "ovl" to login.

BUG=b:178991919
TEST=Click "Goofy" link in overlord UI can open client's goofy UI

Change-Id: Ie0ad85028197a2886ac84873ef2960ab6fb60c42
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/2666634
Reviewed-by: Cheng Yueh <cyueh@chromium.org>
Tested-by: Yilin Yang (kerker) <kerker@chromium.org>
Commit-Queue: Yilin Yang (kerker) <kerker@chromium.org>