pytests: Add brightness to camera test
Brightness test can check the maximum brightness of the camera is
between a specific range. This could be used to test if the camera is
recording black screen or not.

Also refactor the function `CaptureTest` to make the flow clearer.

BUG=b:178328303
TEST=Run test on dut. Check the test can detect different situation. The
maximum brightness should be less than 10 when the camera is cover by
the case.

Change-Id: I419660e290bf120735ec82ee16c735af8895c9bf
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/2675424
Tested-by: Chung-Sheng Wu <chungsheng@chromium.org>
Commit-Queue: Chung-Sheng Wu <chungsheng@chromium.org>
Reviewed-by: Pin-yen Lin <treapking@chromium.org>