pytests: Add max_delta_threshold to audio_loop test
Add threshold to check the maximum delta of the wave. When the mic is
disabled the wave of the mic is flat. Delta is the difference between
two sample. This could be used to test if the wave is flat (which means
the mic may be disabled).

BUG=b:178328303
TEST=Run audio_loop noise test. Check the value when the mic is
disabled/enabled.

Change-Id: I8afd4f4f57a53140a9c282ffe14a8f669414c7c9
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/2675426
Tested-by: Chung-Sheng Wu <chungsheng@chromium.org>
Commit-Queue: Chung-Sheng Wu <chungsheng@chromium.org>
Reviewed-by: Stimim Chen <stimim@chromium.org>
Reviewed-by: Pin-yen Lin <treapking@chromium.org>