hwid/service: return success status for empty hwid
Previously DutLabel API clients used to pass "dummy_hwid" to get
possible_labels in response.  As the migration to One Platform has been
completed, the status KNOWN_BAD_HWID is not supposed to be a valid
response anymore and API will return BAD_REQUEST to the clients.  This
CL takes empty hwid as an indicator to return a success status with the
possible_labels to the clients.

BUG=b:181105376
TEST={FACTORY_REPO}$ ./deploy/cros_hwid_service.sh test

Change-Id: Id2d03bb573073f9591a145631e880c6f499fbc76
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/2719979
Reviewed-by: Yong Hong <yhong@chromium.org>
Tested-by: Clark Chung <ckclark@chromium.org>
Commit-Queue: Clark Chung <ckclark@chromium.org>