factory: update probe_sim behavior
On the shill side, the lifecycle of the Modem device has changed.
Now when the SIM is removed, the device is destroyed.
When the probe_sim.py is polling the state by running the 'modem status'
script, it runs 'mmcli -L' which returns 'No modems were found' when the
device has disappeared (and no longer any line with a Sim property since we
have no device)

Let's use the modem device disappearance as the proxy for the SIM
removal to speed up the detection rather than waiting for a new device
to appear with an empty SIM (after about 15 seconds).

BUG=b:179113777
TEST=on Boten, run the "Probe Sim" factory test and see no timeout and a
proper delay.

Change-Id: I680d3484cb0bb7f94a13e38c31eff6fd56a66c83
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/2739613
Tested-by: Vincent Palatin <vpalatin@chromium.org>
Reviewed-by: Henry Sun <henrysun@google.com>
Reviewed-by: Stimim Chen <stimim@chromium.org>
Commit-Queue: Vincent Palatin <vpalatin@chromium.org>