check_image_version: Allow update from factory branch image
The version of a branched image is in format 'R\d+-\d+.\d+.\d+' and its
LooseVersion is not comparable with the version of a release image.

Remove the prefix to make it a strict version.

A use case is when we accidentally image the third partition with test
image and we want to re-image it back to release-image.

BUG=b:181089724
TEST=run CheckReleaseImage on DUT

Change-Id: I55dbeba8c6c981e909788caf4adb76ac39236de5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/2741888
Reviewed-by: Pin-yen Lin <treapking@chromium.org>
Commit-Queue: Cheng Yueh <cyueh@chromium.org>
Tested-by: Cheng Yueh <cyueh@chromium.org>