dkps: Fix DKPS for new gpg version and python3
DKPS is not being used for a long time, so it was not working because of
python3 migration and the behavior change of new gpg version and
python3 migration.
This CL makes the following changes to pass the unit tests and fix the
errors:
(1) Latest GnuPG requires passphrases by default, so we have to append
    additional flags to disable the protection.
(2) Latest GnuPG requires `gnupghome` to be a existing directory before
    initialization, so we create the folders first at the unit test.
(3) Fix encoding issues for python3 migration.
(4) Other minor fixes.

BUG=None
TEST=dkps_unittest.py
TEST=Manually try DKPS commands from shell.

Change-Id: If855467247a8243eee38fe62342ec12442de85b0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/2784233
Reviewed-by: Yilin Yang (kerker) <kerker@chromium.org>
Reviewed-by: Pin-yen Lin <treapking@chromium.org>
Tested-by: Pin-yen Lin <treapking@chromium.org>
Commit-Queue: Pin-yen Lin <treapking@chromium.org>