setup: Update cros_payload link in NETBOOT.md
If we put link to a symbolic link file in markdown, it will show a
"Symbolic link to xxx" page on the document page. This CL directly link
cros_payload file to the actual shell script file, so the reader won't
see a page for symbolic link.

BUG=b:181954121
TEST=None

Change-Id: I5bc8e29a4a8b8ccbbe782f686e6ae7a8ed6b45e7
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/2813848
Reviewed-by: Jian-Jia Su <jjsu@chromium.org>
Tested-by: Jian-Jia Su <jjsu@chromium.org>
Commit-Queue: Pin-yen Lin <treapking@chromium.org>