wipe: Include /usr/lib for USE=-split-usr
In trying to upgrade sys-apps/util-linux we found that the libraries for
programs such as umount are moved from /lib to /usr/lib because we're
inheriting the usr_ldscript eclass when we didn't before. We want to
move the libraries anyway, because we have disabled the USE flag
'split-usr' globally. This tool needs to mount /usr/lib now too so that
the libraries are there for the wipe tool to work.

BUG=b:169246663
TEST=tast run $DUT factory.Finalize; with CL:2426844

Change-Id: I24318f170015e3e3d8191bcd6a247d66fa2da3ba
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/factory/+/2836630
Tested-by: Stephen Boyd <swboyd@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Stimim Chen <stimim@chromium.org>
Auto-Submit: Stephen Boyd <swboyd@chromium.org>
Commit-Queue: Stimim Chen <stimim@chromium.org>